fix: renamed base_objects to core
