+++ warframe/worldstate/common/__init__.py
@@ -1,2 +1,2 @@
 from .types_ import *
-from .base_objects import *
+from .core import *
